# openapi.yaml
openapi: 3.0.3
info:
  title: Dparty APIs
  description: |-
    Dparty
  contact:
    email: chenyunda218@gmail.com
  version: 0.0.10
servers:
  - url: https://localhost/api/v1
...
paths:
...
  /account/session:
    post:
      tags:
        - Account
      description: Create Login session
      operationId: CreateSession
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: 'Account created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
...
components:
  schemas:
...
    CreateSessionRequest:
      description: 'Login request'
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          minLength: 11
          maxLength: 11
        password:
          type: string
          minLength: 10
          maxLength: 128
...
    Session:
      type: object
      required:
        - token
      properties:
        token:
          type: string
...